+++ src/app.d.ts
@@ -40,19 +40,6 @@ interface Credentials {
   password: string
 }
 
-interface GoogleCredentialResponse {
-	credential: string
-	select_by:
-		| 'auto'
-		| 'user'
-		| 'user_1tap'
-		| 'user_2tap'
-		| 'btn'
-		| 'btn_confirm'
-		| 'btn_add_session'
-		| 'btn_confirm_add_session'
-}
-
 interface MessageAddressee {
   email: string
   name?: string